+++ apps/client/src/widgets/docs/model/types.ts
@@ -1,6 +1,7 @@
 export interface DocsSectionItem {
   label: string;
   href: string;
+  icon?: React.ElementType;
   children?: DocsSectionItem[];
 }
